requirements.txt
annotated-types==0.6.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
colorama==0.4.6
confection==0.1.3
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.1
pathy==0.10.2
Pillow==10.0.1
preshed==3.0.9
pydantic==2.4.2
pydantic_core==2.10.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wasabi==1.1.2
weasel==0.3.2
